cleanup and speedup
A bunch of minor little code cleanups and also a heap of optimisations
to find extra speed; such as replacing for-in loops with proper indexed
for loops on arrays.
Also changed the way one() is called to make add() much faster by removing
the intermediary call and _add(). one() now does a custom `this`.